+++ js/constants.js
@@ -1,6 +1,7 @@
 export const SELL = 0;
-export const BUY = 1;
+export const AUCTION = 1;
 
 export const LOGIN = 0;
 export const SUCCESS = 1;
 export const PRODUCTS = 2;
+export const CATEGORIES = 3;