+++ Lib/UI/changelog.md
@@ -1,4 +1,19 @@
 
+# 13-05-2025 :
+- created a `Component` folder
+	- added `Login.js` file
+		- added `login_component` class
+	- added `Menu.js` file
+		- added `menu_component` class
+			- added `toggle_menu` method
+- removed `Component.js` file
+- added `get_offset_width, get_scroll_width, get_client_width, get_offset_height, get_scroll_height, get_client_height, has_class` methods to `element` class
+- removed `get_width, get_height` methods from `element` class
+
+# 12-05-2025 :
+- added `height, get_height, swap_classes, clone, click` methods to `element` class
+
+
 # 05-05-2025 :
 - added `ws` attribute and `send` method to `WebSocket` class and changed `connect` method to return `this` instead of `ws`